requirements.txt
streamlit
openai
langchain
langchain-google-vertexai
PIL
opencv-python
scikit-image
langchain-community
langchain-openai
python-dotenv
altair==4
watchdog
streamlit_autorefresh
diffusers
accelerate
transformers
langchain-openai
torch
protobuf==3.20.0
PIL
horde-sdk


